requirements.txt
asgiref==3.8.1
certifi==2024.7.4
Django==5.1
django-cors-headers==4.4.0
django-elasticsearch-dsl==8.0
django-environ==0.11.2
djangorestframework==3.15.2
drf-yasg==1.21.7
elastic-transport==8.15.0
elasticsearch==8.15.0
elasticsearch-dsl==8.15.1
exceptiongroup==1.2.2
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.3.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
